Returning response message from cloudflare API and display

+++ src/Resources/js/public.js
@@ -12,12 +12,17 @@ document.addEventListener('DOMContentLoaded', () => {
     let cfoMarkets = document.getElementById('cfo-purge-container').children;
     for (const cfoMarketButton of cfoMarkets) {
         cfoMarketButton.onclick = function () {
-            purgeMarket(cfoMarketButton.dataset.cfoMarket);
+            if (confirm("Are you sure you want to clear the cache for " + cfoMarketButton.children[0].innerHTML)) {
+                purgeMarket(cfoMarketButton.dataset.cfoMarket);
+            }
         };
     }
     
     async function purgeMarket(marketID) {
         try {
+            const resultContainer = document.getElementById('cfo-purge-result');
+            resultContainer.innerHTML = '';
+            resultContainer.classList.add('cfo-loading');
             const response = await fetch('/wp-json/cfo/purge_market?' + new URLSearchParams({market_id: marketID}).toString() , {
                 headers: {
                     'X-WP-Nonce': myUserData.nonce
@@ -25,12 +30,23 @@ document.addEventListener('DOMContentLoaded', () => {
                 credentials: 'same-origin',
             });
             if (!response.ok) {
+                resultContainer.classList.remove('cfo-loading');
                 throw new Error(`Response status: ${response.status}`);
             }
-            const text = await response.text();
-            if (text.length !== 0) {
-                console.log(text.length);
+            const json = await response.json();
+            let textToDisplay = 'Nothing to display.';
+            try {
+                let jsonResponse = JSON.parse(json);
+                if (jsonResponse.success == true) {
+                    textToDisplay = 'Successfully purged.';
+                } else {
+                    textToDisplay = JSON.stringify(jsonResponse.errors);
+                }
+            } catch (error) {
+                textToDisplay = json;
             }
+            resultContainer.classList.remove('cfo-loading');
+            resultContainer.innerHTML = textToDisplay;
         } catch (error) {
             console.error(error);
         }
